requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.10
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pypdf>=3.17.0
huggingface-hub>=0.16.4
llama-cpp-python>=0.2.23
